#24 軽めのスタイル調整+リファクタリング
