+++ frontend/src/plugins/ImagePlugin/index.tsx
@@ -1,16 +1,28 @@
 import { INSERT_IMAGE_COMMAND } from '@/components/Posts/Card/Editor/Command/Image';
 import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
-import { FC } from 'react';
+import { FC, useRef } from 'react';
+import ImageIcon from '@mui/icons-material/Image';
+import { Button } from '@/components/ui/button';
 
 export const ImageItem: FC = () => {
+  const fileInputRef = useRef<HTMLInputElement>(null);
   const [editor] = useLexicalComposerContext();
+  const handleButtonClick = () => {
+    if (fileInputRef.current) {
+      fileInputRef.current.click();
+    }
+  };
   return (
-    <Label>
-      <span>画像を挿入</span>
+    <>
+      <Button type='button' variant={'outline'} onClick={handleButtonClick}>
+        <ImageIcon />
+      </Button>
+
       <Input
         type='file'
+        ref={fileInputRef}
+        className='hidden'
         onChange={async (e: React.ChangeEvent<HTMLInputElement>) => {
           if (e.target.files && e.target.files.length > 0) {
             const file = e.target.files[0];
@@ -29,6 +41,6 @@ export const ImageItem: FC = () => {
           }
         }}
       />
-    </Label>
+    </>
   );
 };
