+++ frontend/src/plugins/LinkToolbarPlugin/index.tsx
@@ -4,6 +4,8 @@ import { TOGGLE_LINK_COMMAND } from '@lexical/link';
 import { Popover, PopoverContent, PopoverTrigger } from '@/components/ui/popover';
 import { useLexicalComposerContext } from '@lexical/react/LexicalComposerContext';
 import { FC, useState } from 'react';
+import LinkIcon from '@mui/icons-material/Link';
+import { Input } from '@/components/ui/input';
 
 export const LinkToolbarItem: FC = () => {
   const [url, setUrl] = useState('');
@@ -12,10 +14,12 @@ export const LinkToolbarItem: FC = () => {
   return (
     <Popover>
       <PopoverTrigger asChild>
-        <Button variant={'outline'}>リンクへ変換</Button>
+        <Button variant={'outline'}>
+          <LinkIcon />
+        </Button>
       </PopoverTrigger>
-      <PopoverContent className='w-80'>
-        <input type='text' value={url} onChange={(e) => setUrl(e.target.value)} placeholder='Enter a URL' />
+      <PopoverContent className='w-80 flex items-center gap-x-4'>
+        <Input type='text' value={url} onChange={(e) => setUrl(e.target.value)} placeholder='Enter a URL' />
         <Button
           onClick={() => {
             if (validateUrl(url)) {
@@ -25,8 +29,9 @@ export const LinkToolbarItem: FC = () => {
             }
             close();
           }}
+          variant={'outline'}
         >
-          リンクに変換
+          <LinkIcon />
         </Button>
       </PopoverContent>
     </Popover>
